agent-os/pyproject.toml
[project]
name = "ia-prende"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ddgs>=9.12.1",
    "groq>=1.1.2",
    "python-dotenv>=1.2.2",
]
